Move default CustomUIOptions implementation to Configurations folder
